option,helm: Define flags for high scale ipcache mode
This agent flag will be used in follow up commits.

This new mode doesn't support the egress gateway for now because the
egress gateway relies on identities from the ipcache to know if a
destination is outside the cluster and if egress policies should
therefore apply. IPsec is also not supported as the key SPIs are
currently stored in the ipcache.

This new mode requires tunneling to be enabled (pod-to-pod traffic will
be sent over the tunnel) and IPv6 to be disabled (our tunnels are only
on IPv4). The IPv4 native routing CIDR must be set because we rely on
that to identify pod-to-pod traffic and encapsulate it.

Signed-off-by: Paul Chaignon <paul@cilium.io>