bpf, config: Encapsulate traffic with pod IPs
In case the high-scale ipcache mode is enabled, we want to encapsulate
pod-to-pod traffic with the pod IP address. The goal of this
encapsulation is simply to carry the security identity for the source.

To know if the destination is a remote pod, we can use the native
routing CIDR. If the destination is not in the ipcache but it is part of
the native routing CIDR, then it's assumed to be a remote pod.

We also need to take this into account when computing the MTU. Thus, the
TunnelExists() function is updated to return true when the high-scale
ipcache mode is enabled.

Signed-off-by: Paul Chaignon <paul@cilium.io>