bpf: Decapsulate traffic encapsulated with pod IPs
When the high-scale ipcache is enabled, we will receive traffic
encapsulated with the pod IP addresses on the native device. Since the
destination IP is not assigned to the host (but to a container), the
Linux stack won't demultiplex it to the overlay device (e.g.,
cilium_vxlan). Instead, the packets will follow their normal way to the
container, via cilium_host if endpoint routes are disabled.

We want to decapsulate the packet before they reach the lxc devices.
We could decapsulate in bpf_lxc, but then the packet paths would be
assymetric. This commit adds support for decapsulation in cilium_host.
Note that will only work when endpoint routes are disabled.

Therefore, in bpf_host, we filter all incoming VXLAN traffic based on
the UDP port and remove the first IP, UDP, VXLAN, and Ethernet headers.
We also parse the source security identity from the VXLAN header. We
then redirect the packet to its expected path, via cilium_vxlan, with
the security identity in skb->cb.

In bpf_overlay, we need a special case for the high-scale ipcache, to
retrieve the security identity from skb->cb instead of getting it from
the tunnel metadata as usual (since packet is already decapsulated at
this point).

Signed-off-by: Paul Chaignon <paul@cilium.io>