bpf: Use world CIDR map
Previous commits introduced a new CRD and its corresponding map. They
tell us which IP addresses belong to entities outside of the cluster, or
at least entities to which we shouldn't encapsulate traffic when in
high-scale IPcache mode.

This commit makes use of the new map. We perform an LPM lookup into the
map to know if we should encapsulate traffic or not.

Signed-off-by: Paul Chaignon <paul@cilium.io>