bpf: Support GENEVE with high-scale ipcache mode
In high-scale ipcache mode, we have a special tunnel where the outer IP
addresses are also the inner IP addresses. For that reason, we need to
manually (using BPF helpers) decapsulate the ingressing packets or the
Linux kernel would deliver encapsulated packets to the target pods.

This manual decapsulation is only performed for VXLAN now. The present
commit adds support for decapsulating GENEVE packets.

Signed-off-by: Paul Chaignon <paul@cilium.io>