bpf, daemon: Move decapsulation for HS ipcache to native device
This commit moves the decapsulation of the special tunnel for high-scale
IPcache mode to the native device. We can't rely on cilium_host to
decapsulate because we won't be going through cilium_host when endpoint
routes are enabled.

As a consequence, native devices are required for Cilium to function in
this mode. Therefore, Cilium agents will attempt to auto-detect the
devices and fatal if they can't.

Signed-off-by: Paul Chaignon <paul@cilium.io>