daemon: Fix use of GetCachedPod with high-scale IPcache
Commit bdb1b6308 ("daemon: Introduce endpointMetadataFetcher")
introduced an alternative implementation to GetCachedPod for when the
high-scale IPcache is enabled and the pod watcher and pod store are
disabled.

There are however a few places that were missed and may still rely on
the GetCachedPod method when high-scale IPcache is enabled. This commit
replaces them with the Fetch method (which abstracts both
implementations).

Fixes: bdb1b6308 ("daemon: Introduce endpointMetadataFetcher")
Signed-off-by: Paul Chaignon <paul@cilium.io>