watcher: Always disable pod watcher with high-scale IPcache
Commit 309edd259 disabled the pod watcher when high-scale IPcache is
enabled, but because of where the skip logic was added, the pod watcher
isn't disabled when disable-endpoint-crd=false. That flag is false by
default so the pod watcher wasn't disabled by default :-(

This commit fixes it by moving the skip logic before the
disable-endpoint-crd check. We also need to signal to the wait group
that we are done or the agent will be blocked on startup.

Fixes: 309edd259 ("k8s/watchers: Prevent creation of pod watcher with high-scale ipcache")
Signed-off-by: Paul Chaignon <paul@cilium.io>