bpf: Don't encapsulate traffic to local pods in HS ipcache mode
In high-scale IPcache mode, pods are excluded from the IPcache and we
therefore can't rely on the IPcache to know when to encapsulate traffic.
We decide to encapsulate or not based on a user-provided set of CIDRs.

Of course, we never want to encapsulate traffic going to local pods. If
endpoint routes are disabled, that is already the case: we will never go
through the encapsulation logic as we will directly deliver packets to
their destination endpoint with a bpf_redirect.

If endpoint routes are enabled however, we need to check if the
destination is a local endpoint, in which case we don't want to
encapsulate. We do that in this commit using the local endpoint map
which contains only endpoints hosted on the node.

Signed-off-by: Paul Chaignon <paul@cilium.io>